style: improve readability

+++ utility/include/io.hpp
@@ -20,13 +20,13 @@ namespace io
 {
 extern const char* version() noexcept;
 
-extern std::string executeCommand(const std::string_view command, const std::uint32_t timeout = 0);
+extern std::string executeCommand(const std::string_view command, const std::size_t timeout = 0);
 extern void waitForUserInput(const std::function<bool(const std::string_view)>& operation, const int timeout = -1);
 extern std::list<std::string> getFileContents(
     const std::string_view filename,
     const bool toLock,
     const bool toReverse = false,
-    const std::uint64_t totalRows = std::numeric_limits<std::uint64_t>::max());
+    const std::size_t totalRows = std::numeric_limits<std::size_t>::max());
 
 //! @brief Custom stream buffer for file descriptors.
 class FDStreamBuffer : public std::streambuf
